requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap4==2024.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-crispy-forms==2.1
django-humanize==0.1.2
django-import-export==3.3.8
djangorestframework==3.15.1
et-xmlfile==1.1.0
ethiopian-date==1.0
future==1.0.0
humanize==4.9.0
import-export==0.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
pkg_resources==0.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzk==0.9
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.11.0
xlrd==2.0.1
xlwt==1.3.0
